amd64 port: Phase 6 — fix LP64 ABI issues and complete distribution build
Fix all remaining build failures needed to produce a successful
`distribution` build for amd64 (exit code 0).

LP64 pointer-size fixes:
- tty/pty: add tty_outkbuf pointer field to store kernel buffer addresses,
  replacing truncated cp_grant_id_t casts to void* on 64-bit
- pty.c, console.c, rs232.c (i386/earm): use tty_outkbuf instead of
  casting 32-bit grant id to pointer
- vnd.c, memory.c, fbd.c, mmcblk.c: change bdr_transfer return type
  from int to ssize_t to match blockdriver callback signature
- vnd.c: fix iov_grant cast via (uintptr_t) intermediate; fix %zd
  format specifiers for ssize_t
- memory.c: change local variable 'a' from u32_t to vir_bytes (8 bytes
  on amd64 vs 4 bytes on i386)

Architecture guard extensions (#if __i386__ → || __x86_64__):
- minix/include/minix/drivers.h: expose bios.h and ports.h for x86_64
  (identical x86 hardware — TIMER_MODE, PORT_B, VDU_* constants needed)
- minix/drivers/tty/tty/arch/i386/rs232.c: expose RS232 port struct
  fields for x86_64 (8250 UART is identical on both platforms)

Build infrastructure:
- memory/Makefile: map MACHINE_CPU=x86_64 to OBJCOPY_ARCH=i386:x86-64
  for objcopy -B flag (x86_64 is not a valid BFD architecture name)
- etc/mtree/NetBSD.dist.x86_64: add usr/libdata/ldscripts directory
  (required for ld linker script installation on 64-bit)
- etc/mtree/NetBSD.dist.tests: add usr/tests/lib/libc/isc directory

+++ minix/commands/vol/vol.c
@@ -273,7 +273,7 @@ int fd;
 		exit(1);
 	}
   } else {
-	if (mtget.mt_blksiz > SSIZE_MAX) {
+	if ((ssize_t)mtget.mt_blksiz > SSIZE_MAX) {
 		fprintf(stderr,
 		"vol: %s: tape block size (%lu) is too large to handle\n",
 			name, (unsigned long) mtget.mt_blksiz);
