+++ minix/lib/libminixfs/cache.c
@@ -336,7 +336,7 @@ static int get_block_ino(struct buf **bpp, dev_t dev, block64_t block, int how,
   if((ino_off % fs_block_size)) {
 
 	printf("cache: unaligned lmfs_get_block_ino ino_off %llu\n",
-		ino_off);
+		(unsigned long long)ino_off);
   	util_stacktrace();
   }
 
@@ -580,7 +580,9 @@ static void put_block(struct buf *bp, int put_flags)
 			printf("libminixfs: no memory for cache block!\n");
 		} else {
 			panic("libminixfs: setblock of %p dev 0x%llx off "
-				"0x%llx failed\n", bp->data, dev, dev_off);
+				"0x%llx failed\n", bp->data,
+				(unsigned long long)dev,
+				(unsigned long long)dev_off);
 		}
 	}
   }
